+++ src/nfc-lib/lib-hw/hw-dev/src/main/include/hw/Device.h
@@ -81,9 +81,9 @@ class Device
 
       virtual bool isStreaming() const = 0;
 
-      virtual int read(BufferType &buffer) = 0;
+      virtual long read(BufferType &buffer) = 0;
 
-      virtual int write(const BufferType &buffer) = 0;
+      virtual long write(const BufferType &buffer) = 0;
 };
 
 }
