+++ src/nfc-lib/lib-hw/hw-radio/src/main/include/hw/radio/MiriDevice.h
@@ -67,9 +67,9 @@ class MiriDevice : public RadioDevice
 
       bool isStreaming() const override;
 
-      int read(SignalBuffer &buffer) override;
+      long read(SignalBuffer &buffer) override;
 
-      int write(const SignalBuffer &buffer) override;
+      long write(const SignalBuffer &buffer) override;
 
       static std::vector<std::string> enumerate();
